Move webhook logic from default handler to postgresql
