+++ pkg/controller/webhooks/default_handler.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/blang/semver/v4"
 	"github.com/go-logr/logr"
 	vshnv1 "github.com/vshn/appcat/v4/apis/vshn/v1"
 	"github.com/vshn/appcat/v4/pkg/common/quotas"
@@ -634,25 +633,4 @@ func checkManualVersionManagementWarnings(maintenance vshnv1.VSHNDBaaSMaintenanc
 				". This is strongly discouraged and may leave your instance without security patches.")
 	}
 	return warnings
-}
-
-// validateNoDowngrade returns an error if newVersion is lower than oldVersion.
-// Both versions are parsed tolerantly, so plain major versions ("15"), semver ("15.9"), and full versions ("15.9.1") are all accepted.
-// If either version is empty or unparseable as an old version, the check is skipped.
-func validateNoDowngrade(oldVersion, newVersion string, path *field.Path) *field.Error {
-	if oldVersion == "" || newVersion == "" {
-		return nil
-	}
-	oldV, err := semver.ParseTolerant(oldVersion)
-	if err != nil {
-		return nil
-	}
-	newV, err := semver.ParseTolerant(newVersion)
-	if err != nil {
-		return field.Invalid(path, newVersion, fmt.Sprintf("invalid version %q", newVersion))
-	}
-	if newV.LT(oldV) {
-		return field.Invalid(path, newVersion, fmt.Sprintf("downgrading from %q to %q is not supported", oldVersion, newVersion))
-	}
-	return nil
 }
