+++ pkg/controller/webhooks/nextcloud.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	valid "github.com/asaskevich/govalidator"
+	"github.com/blang/semver/v4"
 	vshnv1 "github.com/vshn/appcat/v4/apis/vshn/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -113,7 +114,7 @@ func (n *NextcloudWebhookHandler) ValidateUpdate(ctx context.Context, oldObj, ne
 	}
 
 	if nx.Spec.Parameters.Maintenance.PinImageTag == "" {
-		if err := validateNoDowngrade(oldNx.Spec.Parameters.Service.Version, nx.Spec.Parameters.Service.Version, field.NewPath("spec", "parameters", "service", "version")); err != nil {
+		if err := validateNoVersionDowngrade(oldNx.Spec.Parameters.Service.Version, nx.Spec.Parameters.Service.Version); err != nil {
 			allErrs.Add(err)
 		}
 	}
@@ -146,6 +147,32 @@ func (n *NextcloudWebhookHandler) ValidateUpdate(ctx context.Context, oldObj, ne
 	}
 
 	return nil, allErrs.Get()
+}
+
+func validateNoVersionDowngrade(oldVersion, newVersion string) *field.Error {
+	if oldVersion == "" || newVersion == "" {
+		return nil
+	}
+	oldV, err := semver.ParseTolerant(oldVersion)
+	if err != nil {
+		return nil
+	}
+	newV, err := semver.ParseTolerant(newVersion)
+	if err != nil {
+		return field.Invalid(
+			field.NewPath("spec", "parameters", "service", "version"),
+			newVersion,
+			fmt.Sprintf("invalid version %q", newVersion),
+		)
+	}
+	if newV.LT(oldV) {
+		return field.Invalid(
+			field.NewPath("spec", "parameters", "service", "version"),
+			newVersion,
+			fmt.Sprintf("downgrading from %q to %q is not supported; set spec.parameters.maintenance.pinImageTag to override", oldVersion, newVersion),
+		)
+	}
+	return nil
 }
 
 func validateFQDNs(fqdns []string) *field.Error {
